container/pyproject.toml
[project]
name = "hpe-automation-container"
version = "0.1.0"
description = "Automation container for healthy-person-emulator.org"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "Pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "tweepy>=4.14.0",
    "atproto>=0.0.55",
    "Misskey.py>=4.1.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "dlt[bigquery]>=0.4.8",
    "google-cloud-bigquery>=3.10.0",
    "google-auth>=2.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
